Moves code of "Add Tag" command into extension for clarity
