+++ piwigo/Image/Parameters/Tags/TagsViewController.swift
@@ -16,7 +16,7 @@ protocol TagsViewControllerDelegate: NSObjectProtocol {
     func didSelectTags(_ selectedTags: [Tag])
 }
 
-class TagsViewController: UITableViewController, UITextFieldDelegate {
+class TagsViewController: UITableViewController {
 
     weak var delegate: TagsViewControllerDelegate?
 
@@ -37,7 +37,7 @@ class TagsViewController: UITableViewController, UITextFieldDelegate {
      The TagsProvider that fetches tag data, saves it to Core Data,
      and serves it to this table view.
      */
-    private lazy var tagsProvider: TagsProvider = {
+    lazy var tagsProvider: TagsProvider = {
         let provider : TagsProvider = TagsProvider()
         provider.fetchedResultsControllerDelegate = self
         return provider
@@ -71,8 +71,8 @@ class TagsViewController: UITableViewController, UITextFieldDelegate {
         }
     }
 
+    var addAction: UIAlertAction?
     private var addBarButton: UIBarButtonItem?
-    private var addAction: UIAlertAction?
     private var hudViewController: UIViewController?
 
     override func viewDidLoad() {
@@ -277,109 +277,6 @@ class TagsViewController: UITableViewController, UITextFieldDelegate {
         // Update section index
         rebuildABCindex()
         self.tableView.reloadSectionIndexTitles()
-    }
-
-
-    // MARK: - Add tag (for admins only)
-    @objc func requestNewTagName() {
-        let alert = UIAlertController(title: NSLocalizedString("tagsAdd_title", comment: "Add Tag"), message: NSLocalizedString("tagsAdd_message", comment: "Enter a name for this new tag"), preferredStyle: .alert)
-
-        alert.addTextField(configurationHandler: { textField in
-            textField.placeholder = NSLocalizedString("tagsAdd_placeholder", comment: "New tag")
-            textField.clearButtonMode = .always
-            textField.keyboardType = .default
-            textField.keyboardAppearance = AppVars.shared.isDarkPaletteActive ? .dark : .default
-            textField.autocapitalizationType = .sentences
-            textField.autocorrectionType = .yes
-            textField.returnKeyType = .continue
-            textField.delegate = self
-        })
-
-        let cancelAction = UIAlertAction(title: NSLocalizedString("alertCancelButton", comment: "Cancel"), style: .cancel, handler: { action in
-            })
-
-        addAction = UIAlertAction(title: NSLocalizedString("alertAddButton", comment: "Add"), style: .default, handler: { action in
-            // Rename album if possible
-            if (alert.textFields?.first?.text?.count ?? 0) > 0 {
-                self.addTag(withName: alert.textFields?.first?.text)
-            }
-        })
-
-        alert.addAction(cancelAction)
-        if let addAction = addAction {
-            alert.addAction(addAction)
-        }
-        alert.view.tintColor = .piwigoColorOrange()
-        if #available(iOS 13.0, *) {
-            alert.overrideUserInterfaceStyle = AppVars.shared.isDarkPaletteActive ? .dark : .light
-        } else {
-            // Fallback on earlier versions
-        }
-        present(alert, animated: true) {
-            // Bugfix: iOS9 - Tint not fully Applied without Reapplying
-            alert.view.tintColor = .piwigoColorOrange()
-        }
-    }
-
-    func addTag(withName tagName: String?) {
-        // Check tag name
-        guard let tagName = tagName, tagName.count != 0 else {
-            return
-        }
-        
-        // Display HUD during the update
-        showPiwigoHUD(withTitle: NSLocalizedString("tagsAddHUD_label", comment: "Creating Tag…"))
-
-        // Add new tag
-        DispatchQueue.global(qos: .userInteractive).async {
-            self.tagsProvider.addTag(with: tagName, completionHandler: { error in
-                guard let error = error else {
-                    self.updatePiwigoHUDwithSuccess {
-                        self.hidePiwigoHUD(afterDelay: kDelayPiwigoHUD, completion: {})
-                    }
-                    return
-                }
-                self.hidePiwigoHUD {
-                    self.dismissPiwigoError(withTitle: NSLocalizedString("tagsAddError_title", comment: "Create Fail"),
-                                            message: NSLocalizedString("tagsAddError_message", comment: "Failed to…"),
-                                            errorMessage: error.localizedDescription, completion: { })
-                }
-            })
-        }
-    }
-
-
-    // MARK: - UITextField Delegate Methods
-    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
-        // Disable Add/Delete Category action
-        addAction?.isEnabled = false
-        return true
-    }
-
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        // Enable Add/Delete Tag action if text field not empty
-        let finalString = (textField.text as NSString?)?.replacingCharacters(in: range, with: string)
-        var existTagWithName = false
-        if let _ = selectedTags.firstIndex(where: {$0.tagName == finalString}),
-            let _ = nonSelectedTags.firstIndex(where: {$0.tagName == finalString}) {
-            existTagWithName = true
-        }
-        addAction?.isEnabled = (((finalString?.count ?? 0) >= 1) && !existTagWithName)
-        return true
-    }
-
-    func textFieldShouldClear(_ textField: UITextField) -> Bool {
-        // Disable Add/Delete Category action
-        addAction?.isEnabled = false
-        return true
-    }
-
-    func textFieldShouldEndEditing(_ textField: UITextField) -> Bool {
-        return true
-    }
-
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        return true
     }
 }
